finish stellar sign transaction, send signed transaction

+++ NRLWalletSDK Example/TransactionViewController.swift
@@ -17,14 +17,14 @@ internal class TransactionViewController: UIViewController {
     
     @IBAction func OnSend() {
         let to = tfTo.text;
-        let value = UInt64(tfValue.text!)
-        let fee = UInt64(tfFee.text!)
+        let value = Double(tfValue.text!)
+        let fee = Double(tfFee.text!)
         
         guard let wallet = coinWallet else {
             print("setStellarWallet Error: cannot init wallet!")
             return
         }
-        
+        //GD6E7XNB4W2KEIFN3MVN22FS5GCBTXI44USLCT6FWBMLLHENOQIYZAYD
         wallet.sendTransaction(to: to!, value: value!, fee: fee!) { (err, tx) -> () in
             switch (err) {
             case NRLWalletSDKError.nrlSuccess:
